Update internal unit MaxHP. Renamed HealthPoints to HitPoints. Spacing. Updated unit struct. Hopefully didn't forget to count and mess up alignment. Changed HitPoints to s32 instead of u32. UnitImpl::useTech now supports Archon Warp and Dark Archon Meld (issue 189). Added ShiftSelectSingle to OrderTypes.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@2103 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BW/UnitType.h
@@ -50,7 +50,7 @@ namespace BW
       u8                                   requiredTech() const;
       u8                                   armorUpgrade() const;
 
-      u16                  maxHitPoints() const;
+      s32                  maxHitPoints() const;
       u16                  maxShields() const;
       u16                  maxEnergy() const;
       u8                   armor() const;
